Home Accessory Architect v12.0.0 Merlin

+++ libs/adv_hlw/adv_hlw.h
@@ -14,9 +14,9 @@ extern "C" {
 
 int adv_hlw_unit_create(const int8_t gpio_cf, const int8_t gpio_cf1, const int8_t gpio_sel, const unsigned int current_mode, const unsigned int interrupt_type);
 
-double adv_hlw_get_voltage_freq(const int gpio);
-double adv_hlw_get_current_freq(const int gpio);
-double adv_hlw_get_power_freq(const int gpio);
+float adv_hlw_get_voltage_freq(const int gpio);
+float adv_hlw_get_current_freq(const int gpio);
+float adv_hlw_get_power_freq(const int gpio);
 
 #ifdef __cplusplus
 }
